Remove manual code for PrintCapabilities
