+++ gdk4-wayland/src/auto/versions.txt
@@ -1,2 +1,2 @@
-Generated by gir (https://github.com/gtk-rs/gir @ 4395f43)
+Generated by gir (https://github.com/gtk-rs/gir @ b5d88a6)
 from gir-files (https://github.com/gtk-rs/gir-files.git @ 7d95377)
